finished food and meat collection and food requests

+++ app/Http/Controllers/JobcardsController.php
@@ -2,9 +2,11 @@
 
 namespace App\Http\Controllers;
 
+use App\Imports\JobcardImport;
 use App\Models\Jobcard;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
+use Maatwebsite\Excel\Facades\Excel;
 
 class JobcardsController extends Controller
 {
@@ -50,36 +52,12 @@ class JobcardsController extends Controller
         }
         else {
             try {
-                $food = Jobcard::where('card_type','=','food')->get();
-
-                if($food->count() == 0)
+                if ($request->card_type == "food")
                 {
-                    $card = Jobcard::create([
-                        'card_number' => $request->input('card_number'),
-                        'date_opened' => $request->input('date_opened'),
-                        'card_month' => $request->input('card_month'),
-                        'card_type' => $request->input('card_type'),
-                        'quantity' => $request->input('quantity'),
-                        'remaining' => $request->input('quantity'),
-                    ]);
-                    $card->save();
+                    $food = Jobcard::where('card_type','=','food')->get();
 
-                    return redirect('jobcards')->with('success','Jobcard has been created successfully');
-
-                } else {
-                    $i = 0;
-                    foreach ($food as $f)
+                    if($food->count() == 0)
                     {
-                        if ($f->remaining > 0)
-                        {
-                            $i++;
-                        }
-                    }
-
-                    if ($i > 0)
-                    {
-                        return back()->with('error','There is a jobcard with remaining units');
-                    } else {
                         $card = Jobcard::create([
                             'card_number' => $request->input('card_number'),
                             'date_opened' => $request->input('date_opened'),
@@ -90,7 +68,83 @@ class JobcardsController extends Controller
                         ]);
                         $card->save();
 
-                        return redirect('jobcards')->with('success','Job card has been added successfully');
+                        return redirect('jobcards')->with('success','Jobcard has been created successfully');
+
+                    } else {
+                        $i = 0;
+                        foreach ($food as $f)
+                        {
+                            if ($f->remaining > 0)
+                            {
+                                $i++;
+                            }
+                        }
+
+                        if ($i > 0)
+                        {
+                            return back()->with('error','There is a Food jobcard with remaining units');
+
+                        } else {
+                            $card = Jobcard::create([
+                                'card_number' => $request->input('card_number'),
+                                'date_opened' => $request->input('date_opened'),
+                                'card_month' => $request->input('card_month'),
+                                'card_type' => $request->input('card_type'),
+                                'quantity' => $request->input('quantity'),
+                                'remaining' => $request->input('quantity'),
+                            ]);
+                            $card->save();
+
+                            return redirect('jobcards')->with('success','Job card has been added successfully');
+                        }
+                    }
+                }
+
+                if ($request->card_type == "meat")
+                {
+                    $meat = Jobcard::where('card_type','=','meat')->get();
+
+                    if($meat->count() == 0)
+                    {
+                        $card = Jobcard::create([
+                            'card_number' => $request->input('card_number'),
+                            'date_opened' => $request->input('date_opened'),
+                            'card_month' => $request->input('card_month'),
+                            'card_type' => $request->input('card_type'),
+                            'quantity' => $request->input('quantity'),
+                            'remaining' => $request->input('quantity'),
+                        ]);
+                        $card->save();
+
+                        return redirect('jobcards')->with('success','Job Card has been created successfully');
+
+                    } else {
+                        $i = 0;
+                        foreach ($meat as $m)
+                        {
+                            if ($m->remaining > 0)
+                            {
+                                $i++;
+                            }
+                        }
+
+                        if ($i > 0)
+                        {
+                            return back()->with('error','There is a Meat Job Card with remaining units');
+
+                        } else {
+                            $card = Jobcard::create([
+                                'card_number' => $request->input('card_number'),
+                                'date_opened' => $request->input('date_opened'),
+                                'card_month' => $request->input('card_month'),
+                                'card_type' => $request->input('card_type'),
+                                'quantity' => $request->input('quantity'),
+                                'remaining' => $request->input('quantity'),
+                            ]);
+                            $card->save();
+
+                            return redirect('jobcards')->with('success','Job card has been added successfully');
+                        }
                     }
                 }
 
@@ -131,7 +185,64 @@ class JobcardsController extends Controller
      */
     public function update(Request $request, Jobcard $jobcard)
     {
-        //
+        $validator = Validator::make($request->all(),[
+            'card_number' => 'required|unique:jobcards,card_number,'.$jobcard->card_number,
+            'date_opened' => 'required',
+            'card_month' => 'required',
+            'card_type' => 'required',
+            'quantity' => 'required',
+        ]);
+
+        if ($validator->fails())
+        {
+            return redirect()->back()->withErrors($validator)->withInput();
+
+        } else {
+
+            try {
+                // checking the request type
+                if ($request->card_type == "food")
+                {
+                    return back()->with('warning','Please note that food humber job card cannot be edited. ');
+
+                } else {
+
+                    $total = $jobcard->issued + $jobcard->extras_previous;
+
+                    if ($total > 0)
+                    {
+                        // can only edit qty
+                        $current_qty = $jobcard->quantity;
+                        if ($current_qty > $request->quantity)
+                        {
+                            return back()->with('error','Supplied quantity is less than the existing quantity');
+
+                        } else{
+
+                            $jobcard->quantity = $request->quantity;
+                            $jobcard->save();
+
+                            return redirect('jobcards')->with('success','Job card has been updated successfully');
+                        }
+
+                    } else {
+                        // can edit evertthing on the card
+                        $jobcard->card_number = $request->input('card_number');
+                        $jobcard->date_opened = $request->input('date_opened');
+                        $jobcard->card_month = $request->input('card_month');
+                        $jobcard->card_type = $request->input('card_type');
+                        $jobcard->quantity = $request->input('quantity');
+                        $jobcard->save();
+
+                        return redirect('jobcards')->with('success','Jobcard has been updated successfully');
+                    }
+                }
+
+            } catch (\Exception $e) {
+                echo "Error - ".$e;
+            }
+        }
+
     }
 
     /**
@@ -143,12 +254,22 @@ class JobcardsController extends Controller
     public function destroy($id)
     {
         $jobcard = Jobcard::findOrFail($id);
-        $jobcard->delete();
 
-        return redirect('jobcards')->with('success','Jobcard has been deleted successfully');
+        // check if the jobcard has been issued or not
+        $total = $jobcard->issued + $jobcard->extras_previous;
+
+        if ($total == 0)
+        {
+            $jobcard->delete();
+
+            return redirect('jobcards')->with('success','Jobcard has been deleted successfully');
+        } else {
+
+            return back()->with('error','Job Card Cannot be deleted.');
+        }
     }
 
-    public function downloadAllocationForm()
+    public function downloadJobcardForm()
     {
         $myFile = public_path("starter-downloads/jobcards.xlsx");
 
@@ -160,8 +281,19 @@ class JobcardsController extends Controller
         return view('jobcards.import');
     }
 
-    public function uploadJobcards()
+    public function uploadJobcards(Request $request)
     {
-        dd($request);
+        $validator = Validator::make($request->all(),[
+            'jobcard' => 'required',
+
+        ]);
+
+        if ($validator->fails()) {
+            return back()->withErrors($validator)->withInput();
+        }
+
+        Excel::import(new JobcardImport,request()->file('jobcard'));
+
+        return redirect('jobcards')->with('Data has been imported successfully');
     }
 }
